feat: implement authentication context and reducer; refactor AuthProvider and hooks for improved state management
