+++ frontend/src/app/hooks/useAuth.ts
@@ -5,7 +5,6 @@ const Authkey = {
     all: ['user'] as const,
     user: () => [...Authkey.all, 'user'] as const
 }
-
 export const useAuth = () => {
     const quryClinet = useQueryClient();
     const {
