+++ frontend/src/app/types/auth.d.ts
@@ -1,4 +1,4 @@
- type User = {
+type User = {
     id: string;
     name: string;
     email: string;
@@ -10,9 +10,9 @@ export interface LoginCredentials {
     password: string;
 }
 export interface RegisterCredentials {
-    name: "string";
-    email: "string";
-    password: "string";
+    name: string;
+    email: string;
+    password: string;
     role: "ADMIN" | "USER";
 }
 
@@ -26,6 +26,10 @@ export interface AuthContextType {
     user: User | null;
     token: string | null;
     isAuthenticated: boolean;
-    isLoading: boolean;
-    error: string | null;
+    isLoaing: boolean;
+    Error: string | null;
+    login: (email: string, password: string) => Promise<void>;
+    logout: () => void;
+    register: (name: string, email: string, password: string, role: "ADMIN" | "USER") => Promise<void>;
+    clearError: () => void;
 };
